모듈화 추가 진행

+++ raytracing_lib/src/ray.rs
@@ -19,7 +19,7 @@ impl Ray {
         self.dir
     }
 
-    pub fn at(&self, t: f32) -> Vec3 {
+    pub fn at(&self, t: f64) -> Vec3 {
         self.orig + self.dir * t
     }
 }
