+++ src/main.rs
@@ -1,12 +1,12 @@
 use image::ImageBuffer;
-use raytracing_lib::Ray;
-use raytracing_lib::Vec3;
+use raytracing_lib::ray::Ray;
+use raytracing_lib::vec3::Vec3;
 
 fn main() {
     // 이미지 높이를 계산하고 1 이상인지 확인.
     let aspect_ratio = 16.0 / 9.0;
     let image_width = 400;
-    let image_height_aspect = (image_width as f32 / aspect_ratio) as u32;
+    let image_height_aspect = (image_width as f64 / aspect_ratio) as u32;
     let image_height = if image_height_aspect < 1 {
         1
     } else {
@@ -17,7 +17,7 @@ fn main() {
     // 카메라
     let focal_length = 1.0;
     let viewport_height = 2.0;
-    let viewport_width = viewport_height * (image_width as f32 / image_height as f32);
+    let viewport_width = viewport_height * (image_width as f64 / image_height as f64);
     let camera_center = Vec3::new(0.0, 0.0, 0.0);
 
     // 수평 뷰포트 가장자리를 가로질러 벡터를 계산하고, 수직 뷰포트 가장자리를 따라 아래로 벡터를 계산.
@@ -25,8 +25,8 @@ fn main() {
     let viewport_v = Vec3::new(0.0, -viewport_height, 0.0);
 
     // 픽셀에서 픽셀로의 수평 및 수직 델타 벡터를 계산.
-    let pixel_delta_u = viewport_u / image_width as f32;
-    let pixel_delta_v = viewport_v / image_height as f32;
+    let pixel_delta_u = viewport_u / image_width as f64;
+    let pixel_delta_v = viewport_v / image_height as f64;
 
     // 왼쪽 상단 픽셀의 위치를 ​​계산.
     let viewport_upper_left =
@@ -39,7 +39,7 @@ fn main() {
         eprint!("\r남은 스캔라인: {:0height_digit$}", image_height - y);
         for x in 0..image_width {
             let pixel_center =
-                pixel00_loc + (pixel_delta_u * x as f32) + (pixel_delta_v * y as f32);
+                pixel00_loc + (pixel_delta_u * x as f64) + (pixel_delta_v * y as f64);
             let ray_direction = pixel_center - camera_center;
             let r = Ray::new(camera_center, ray_direction);
 
@@ -71,7 +71,7 @@ fn ray_color(r: Ray) -> Vec3 {
     Vec3::new(1.0, 1.0, 1.0) * (1.0 - a) + Vec3::new(0.5, 0.7, 1.0) * a
 }
 
-fn hit_sphere(center: Vec3, radius: f32, r: Ray) -> f32 {
+fn hit_sphere(center: Vec3, radius: f64, r: Ray) -> f64 {
     let oc = r.origin() - center;
     let a = r.direction().length_squared();
     let half_b = oc.dot(r.direction());
